Update gdd.py
This version of gdd.py should work perfectly but I should mention that there are some warning that python interpreter might send. If we ignore them, it is complete by now. 
This file creates a *.gdd file in the /csv_data/ direcotry with the same file name as its input argument.